third_party/WebKit/LayoutTests/editing/style/create-block-for-style-006.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>
Should see "bar" and "baz" centered in their own boxes. The "foo" line should be in its own red box, but not centered.
</p>
<div id="sample" contenteditable>foo<br>bar<br>baz</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    var sample = document.getElementById('sample');

    selection.collapse(sample, 2);
    selection.extend(sample.lastChild, 1);
    document.execCommand('justifyCenter');

    assert_equals(sample.innerHTML, 'foo<br><div style="text-align: center;">bar</div><div style="text-align: center;">baz</div>', 'innerHTML');
    assert_false(selection.isCollapsed, 'isCollapsed');
    assert_equals(selection.anchorNode, sample.childNodes[2].firstChild, 'anchorNode');
    assert_equals(selection.anchorOffset, 0, 'anchorOffset');
    assert_equals(selection.focusNode, sample.childNodes[3].firstChild, 'focusNode');
    assert_equals(selection.focusOffset, 1, 'focusOffset');
});
</script>
